requirements.txt
cadquery>=2.4.0
numpy
scipy
trimesh
flask
requests
torch>=2.0.0
transformers>=4.45.0
peft>=0.17.0
accelerate>=0.34.0
pillow
six
Jinja2
tinydb
rectpack
anytree
safetensors
sentencepiece
cqparts_fasteners
cqparts_misc
# Optional: For quantization (uncomment if needed)
# bitsandbytes

